Ensure the dataset is loaded as float32 instead of float64

+++ specialcouscous/scaling_dataset.py
@@ -284,8 +284,8 @@ def read_scaling_dataset_from_hdf5(
 
     def dataset_from_group(group: h5py.Group) -> SyntheticDataset:
         return SyntheticDataset(
-            group["x"][...],
-            group["y"][...],
+            group["x"][...].astype(np.float32),
+            group["y"][...].astype(np.float32),
             n_samples=int(group.attrs["n_samples"]),
             n_classes=int(n_classes),
         )
@@ -589,7 +589,7 @@ def add_useless_features(
     """
     # Create useless features from random noise
     n_samples = x.shape[0]
-    useless_features = random_state.standard_normal(size=(n_samples, n_useless))
+    useless_features = random_state.standard_normal(size=(n_samples, n_useless)).astype(np.float32)
     # Shift and scale the new features
     useless_features = (useless_features + shift) * scale
